Fix globals plugin for type checking
