--- frontend/types/index.d.ts
@@ -1,7 +1,0 @@
-import { Auth } from "@nuxtjs/auth-next";
-
-declare module "vue/types/vue" {
-  interface Vue {
-    $auth: Auth;
-  }
-}
